Fixed issue #18368: export_statistics from API Yields error
Tests moved to ExportStatisticsTest.php
